// src/Interfaces/JobPostingModel.ts
export default interface JobPosting {
    jobId: number;
    title: string;
    description: string;
    location: string;
    experienceRequired: string;
    skills: string;
    salaryRange: string;
    postedDate: string;
    expiryDate: string;
    jobType: string;
    isActive: boolean;
  }
      
    